starting to migrate the layout
