+++ yarn-dedup.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+yarn yarn-deduplicate -s fewer yarn.lock;
+yarn install # Don't forget this step.
